Fishing Mini-Game Implementation
Began implementation for basic fishing mini-game, allowing developer to choose a list of fish to be able to be spawned and their spawn rate. A list of potential fish is then generated when the player casts their hook .This list is then shuffled, and a skill check is performed on each fish based on its spawn rate value. If it is less than or equal to the fish's value, it will succeed and be hooked on. Otherwise, it will wait a short amount of time and then move on to the next fish.